Integrate Tokio v1.46.1 task spawn locations (#2)
Tokio PRs tokio-rs/tokio#7417 and tokio-rs/tokio#7440 add the source
code locations at which a task was spawned to the `TaskMeta` structure
provided to the runtime task hook callbacks. These will give us the
ability to associate task IDs (which identify a task at runtime) with a
source code location in the instrumented program, which is going to make
everything a lot easier. This PR picks up those changes and updates the
minimum Tokio dependency to 1.46.1.

Closes #1